+++ internal/cli/client/client.go
@@ -37,3 +37,13 @@ func SendRequestAndReadResponse(url *url.URL, enableAuth bool, method string, bo
 	}
 	return data, nil
 }
+
+func CheckConnection(url string) error {
+	_, err := http.Get(url)
+	if err != nil {
+		return fmt.Errorf("\x1b[1;37;41mUNABLE TO CONNECT\x1b[0m | %s\n\t↳ %v",
+			"Did you forget to start the server?",
+			err)
+	}
+	return nil
+}
